Read-Media-Gemini/requirements.txt
python-dotenv
google-genai
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
